requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\WSE App\sabrisangjaya.github.io\twitterbot\newbot\main.py: 10
Pillow == 7.2.0

# D:\WSE App\sabrisangjaya.github.io\twitterbot\newbot\main.py: 11
pandas == 1.1.1

# D:\WSE App\sabrisangjaya.github.io\twitterbot\newbot\main.py: 3
requests == 2.24.0

# D:\WSE App\sabrisangjaya.github.io\twitterbot\newbot\main.py: 4
schedule == 1.1.0

# D:\WSE App\sabrisangjaya.github.io\twitterbot\newbot\main.py: 2
tweepy == 4.4.0

xlrd == 1.1.0
